updated to include the alpha one hot vectors (make sure to rebuild with openvdb updates

+++ examples/cpp/datasets/KITTIOdometry.cpp
@@ -94,12 +94,12 @@ std::vector<Eigen::Vector3d> ReadKITTIVelodyne(const std::string& path) {
     return points;
 }
 
-std::vector<uint16_t> ReadKITTISemantics(const std::string& path) {
+std::vector<int> ReadKITTISemantics(const std::string& path) {
     // Load semantics info from text file
     std::ifstream infile(path.c_str()); // Open the input file
     assert(infile.is_open() && "ReadKITTISemantics| not able to open file");
 
-    std::vector<uint16_t> labels; // Vector to store uint16_t numbers
+    std::vector<int> labels; // Vector to store uint16_t numbers
     std::string line;
 
     while (std::getline(infile,line)) {
@@ -209,9 +209,9 @@ KITTIDataset::KITTIDataset(const std::string& kitti_root_dir,
     label_files_ = GetLabelFiles(fs::absolute(kitti_sequence_dir / "labels_txt/"), n_scans);
 }
 
-std::tuple<std::vector<Eigen::Vector3d>, std::vector<uint16_t>, Eigen::Vector3d> KITTIDataset::operator[](int idx) const {
+std::tuple<std::vector<Eigen::Vector3d>, std::vector<int>, Eigen::Vector3d> KITTIDataset::operator[](int idx) const {
     std::vector<Eigen::Vector3d> points = ReadKITTIVelodyne(scan_files_[idx]);
-    std::vector<uint16_t> semantics = ReadKITTISemantics(label_files_[idx]);
+    std::vector<int> semantics = ReadKITTISemantics(label_files_[idx]);
 
     // if (preprocess_) PreProcessCloud(points, min_range_, max_range_); // if this is enabled, the preprocessing will make the length of the laser scan points shorter than the # of labels
     if (apply_pose_) TransformPoints(points, poses_[idx]);
